refactor cosmos submissions
