+++ utils/utilsEspo.py
@@ -2,6 +2,7 @@ from fastapi import HTTPException, Header
 from utils.logger import logger
 from utils.utilsKobo import update_submission_status
 
+
 def espo_request(submission, espo_client, method, entity, params=None, logs=None):
     """Make a request to EspoCRM. If the request fails, update submission status in CosmosDB."""
     try:
